Exiting a script during execution leaves dirty files all over the project's tree, and that's not desirable, at least for me. I prefer letting the script finish its thing, so that it cleans up any garbage, instead of having to clean it up manually. It's a pain expecially if it fails before unmounting stuff. Also, who on earth doesn't have sudo?

+++ make_me.sh
@@ -1,27 +1,24 @@
-#!/bin/bash -e
+#!/bin/bash
 
-# compiler path
-export CC=${CC:-gcc}
+# path to the cross compiler
+export CC=
 
 printf "== DankOS autobuild tool ==\n\n"
 
-[[ $(type -P $CC) ]] || ( printf "Error: $CC could not be found.\n" ; exit 1 )
-
 if [[ $EUID -ne 0 ]]; then
 printf "This script requires access to root privileges for mounting the image file.\n"
-sudo true
+exit 1
 fi
 
-# Some systems might not have sudo. If we're here without it, we already
-# have root rights, so just set it as an alias for eval.
-
-[[ $(type -P "sudo") ]] || alias sudo=eval
+if [[ "$CC" = "" ]]; then
+printf "C compiler not specified, skipping C sources.\n"
+fi
 
 # Backup data to a "dankos.old" file.
 printf "All data previously stored in 'dankos.img' will be lost (if it exists)!\n"
 printf "A 'dankos.old' backup will be made as a failsafe.\n"
-rm dankos.old 2> /dev/null || true
-mv dankos.img dankos.old 2> /dev/null || true
+rm dankos.old 2> /dev/null
+mv dankos.img dankos.old 2> /dev/null
 
 printf "Assembling bootloader...\n"
 nasm bootloader/bootloader.asm -f bin -o dankos.img
@@ -49,38 +46,42 @@ do
     nasm "$asm_file" -f bin -o "tmp/${base_name}.bin"
 done
 
+if [[ "$CC" != "" ]]; then
+
 printf "Compiling content of the C sources in the 'sources' directory...\n"
 for c_file in sources/*.c
 do
 	base_name=${c_file%.c}
 	base_name=${base_name:8}
 	printf "Compiling '$c_file'...\n"
-	$CC -c -m16 -nostdlib -nostartfiles -nodefaultlibs -fno-builtin "$c_file" -o "tmp/${base_name}.o" -masm=intel -D __DANKOS__
+	$CC -c -m16 -nostdlib -nostartfiles -nodefaultlibs -fno-builtin "$c_file" -o "tmp/${base_name}.o" -masm=intel
 	ld -T linker_script --oformat binary -melf_i386 "tmp/${base_name}.o" -o "tmp/${base_name}.tmp"
-	sed 's/\xC9\xC3/\x66\x67\xC9\xC3/g' "tmp/${base_name}.tmp" > "tmp/${base_name}.bin"			# Bodge for leave instruction
+	# Bodge for leave instruction
+	sed 's/\xC9\xC3/\x66\x67\xC9\xC3/g' "tmp/${base_name}.tmp" > "tmp/${base_name}.bin"
 	rm "tmp/${base_name}.o"
 	rm "tmp/${base_name}.tmp"
 done
 
+fi
+
 printf "Creating mount point for image...\n"
 mkdir mnt
 
-sudo -s eval '
-printf "Mounting image...\n";
-mount dankos.img ./mnt;
+printf "Mounting image...\n"
+mount dankos.img ./mnt
 
-printf "Copying files to image...\n";
-cp -r extra/* mnt/ 2> /dev/null;
-cp tmp/* mnt/;
-cp LICENSE.md mnt/license.txt;
+printf "Copying files to image...\n"
+cp -r extra/* mnt/ 2> /dev/null
+cp tmp/* mnt/
+cp LICENSE.md mnt/license.txt
 
-printf "Unmounting image...\n";
-sync;
-umount ./mnt;'
+printf "Unmounting image...\n"
+sync
+umount ./mnt
 
 printf "Cleaning up...\n"
 rm -rf tmp
-rmdir mnt
+rm -rf mnt
 
 printf "Done!\n\n"
 printf "If everything executed correctly, a file named 'dankos.img'\n"
